Switched UInt to Int and Float to Double to avoid excessive casting

+++ SwiftWizard/Classes/Autocorrelator.swift
@@ -9,7 +9,7 @@ final class Autocorrelator {
     static func getCoefficients(coefficients: [Double], buffer: Buffer) -> [Double] {
         var newCoefficients: [Double] = Array(repeating: 0, count: 11)
         for i in 0...10 {
-            newCoefficients[i] = lagSum(lag: UInt(i), buffer: buffer)
+            newCoefficients[i] = lagSum(lag: i, buffer: buffer)
         }
         return newCoefficients
     }
@@ -19,18 +19,18 @@ final class Autocorrelator {
     }
     
     // Given lag in units time, lagSum returns a "sum of squares" for consecutive lag-separated values
-    private static func lagSum(lag: UInt, buffer: Buffer) -> Double {
-        let samples: UInt   = buffer.size - lag
+    private static func lagSum(lag: Int, buffer: Buffer) -> Double {
+        let samples: Int   = buffer.size - lag
         var sum:     Double = 0.0
         
         for i in 0..<samples {
-            sum += buffer.samples[Int(i)] * buffer.samples[Int(i + lag)]
+            sum += buffer.samples[Int(i)] * buffer.samples[i + lag]
         }
         return sum
     }
     
     // Autocorrelates coefficients while also normalizing to given min/max period values
-    static func normalizedCoefficients(coefficients: [Double], buffer: Buffer, minPeriod: UInt, maxPeriod: UInt) -> [Double]{
+    static func normalizedCoefficients(coefficients: [Double], buffer: Buffer, minPeriod: Int, maxPeriod: Int) -> [Double]{
         var newNormalizedCoefficients: [Double] = coefficients
         for lag in 0...maxPeriod {
             if lag < minPeriod {
@@ -41,14 +41,14 @@ final class Autocorrelator {
             var sumOfSquaresEnd:    Double  = 0.0
             
             var sum:                Double  = 0.0
-            let samples:            UInt    = buffer.size - lag
+            let samples:            Int     = buffer.size - lag
             
             for i in 0..<samples {
-                sum += buffer.samples[Int(i)] * buffer.samples[Int(i + lag)]
-                sumOfSquaresBegin   += pow(buffer.samples[Int(i)], 2)
-                sumOfSquaresEnd     += pow(buffer.samples[Int(i + lag)], 2)
+                sum += buffer.samples[Int(i)] * buffer.samples[i + lag]
+                sumOfSquaresBegin   += pow(buffer.samples[i], 2)
+                sumOfSquaresEnd     += pow(buffer.samples[i + lag], 2)
             }
-            newNormalizedCoefficients[Int(lag)] = sum / sqrt(sumOfSquaresBegin * sumOfSquaresEnd)
+            newNormalizedCoefficients[lag] = sum / sqrt(sumOfSquaresBegin * sumOfSquaresEnd)
         }
         return newNormalizedCoefficients
     }
